Adding dynamic terminal resize
Actually added the ability to send commands that are not
processed as a terminal input. They start with \r~ and end with ~,
which serve as escape chanaters.
The escapes are not part of the command, they only signify that
non-terminal string is coming.

At this moment "size cols rows" is recognized.